新增:messageController資料驗證

+++ app/Http/Controllers/MessageController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Message;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 
 class MessageController extends Controller
 {
@@ -26,6 +27,14 @@ class MessageController extends Controller
      */
     public function store(Request $request)
     {
+        if (trim($request->name) == '') {
+            return new Response('No enter name');
+        }
+
+        if (trim($request->content) == '') {
+            return new Response('No enter content');
+        }
+
         $newMessage = new Message;
         $newMessage->name = $request->input('name');
         $newMessage->content = $request->input('content');
@@ -43,6 +52,11 @@ class MessageController extends Controller
     public function edit($id)
     {
         $message = Message::find($id);
+
+        if (!$message) {
+            return new Response('Not found message');
+        }
+
         return view('messages.edit', ['message' => $message]);
     }
 
@@ -55,7 +69,16 @@ class MessageController extends Controller
      */
     public function update(Request $request)
     {
+        if (trim($request->content) == '') {
+            return new Response('No enter content');
+        }
+
         $updateMessage = Message::find($request->input('id'));
+
+        if (!$updateMessage) {
+            return new Response('Not found message');
+        }
+
         $updateMessage->content = $request->input('content');
         $updateMessage->save();
 
@@ -70,8 +93,12 @@ class MessageController extends Controller
      */
     public function destroy(Request $request)
     {
-        $id = $request->id;
-        $deleteMessage = Message::find($id);
+        $deleteMessage = Message::find($request->id);
+
+        if (!$deleteMessage) {
+            return new Response('Not found message');
+        }
+
         $deleteMessage->delete();
 
         return redirect(route('messages.index'));
